+++ src/types.ts
@@ -1,3 +1,9 @@
+export type ProviderConfig = {
+  nextjs?: {
+    useRouterRefreshOnSave?: boolean;
+  };
+};
+
 export type ComponentDetails = {
   id: string;
   name: string;